Fixed shortest paths

+++ Cities/metropolis.jl
@@ -21,7 +21,7 @@ longTrvlCost(trvltime) = trvltime * driveCost/120
 
 type Metropolis <: TaxiProblem
   network::Network
-  roadTime::SparseMatrixCSC{Float64, Int}
+  roadTime::SparseMatrixCSC{Int, Int}
   roadCost::SparseMatrixCSC{Float64, Int}
   custs::Array{Customer,1}
   taxis::Array{Taxi,1}
